+++ HiveShock.Avalonia/Views/Pages/DiagnosticPage.axaml.cs
@@ -1,0 +1,11 @@
+using Avalonia.Controls;
+
+namespace HiveShock.Avalonia.Views.Pages;
+
+public partial class DiagnosticPage : UserControl
+{
+    public DiagnosticPage()
+    {
+        InitializeComponent();
+    }
+}